python-server/requirements-activation.txt
# Requirements for activation code generation script
supabase>=2.0.0
qrcode[pil]>=7.4.2
reportlab>=4.0.0
python-dotenv>=1.0.0
